requirements.txt
# https://raw.githubusercontent.com/apache/airflow/constraints-2.1.4/constraints-3.8.txt

apache-airflow==2.7.1
apache-airflow-providers-amazon==4.0.0
apache-airflow-providers-cncf-kubernetes==4.1.0
apache-airflow-providers-slack==5.0.0
apache-airflow-providers-slack[http]==5.0.0
apache-airflow-providers-postgres==5.0.0
boto3==1.24.23
botocore==1.27.96
kubernetes==23.6.0
numpy==1.21.6
requests~=2.32.0
PyYAML~=6.0
pylint==2.17.7
pylint-airflow==0.1.0a1
openpyxl==3.1.4
opensearch-py==2.8.0
pandas~=1.5.3
typing_extensions~=4.4.0
psycopg2~=2.9.9
pendulum~=2.1.2